# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "dikep15@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/v1/articles": {
            "get": {
                "summary": "Find all the articles",
                "description": "Returns all the articles",
                "operationId": "2bb6d18420d6929d1d4175ace6521d21",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "articles not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "//api/v1/articles/{id}": {
            "get": {
                "tags": [
                    "articles"
                ],
                "summary": "Find articles by ID",
                "description": "Returns a single articles",
                "operationId": "56175c3e02d9af5eb71fcefb5aaa5422",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of articles to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "articles not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/articles/{id}/comment": {
            "get": {
                "tags": [
                    "articles"
                ],
                "summary": "Find articles by ID with their comments",
                "description": "Returns a single articles with comments",
                "operationId": "d9b56681f4ff3bb2091b341ee5bb5064",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of articles to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "articles not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/articles/{id}/like": {
            "get": {
                "tags": [
                    "articles"
                ],
                "summary": "Find articles by ID with their likes",
                "description": "Returns a single articles with likes",
                "operationId": "b4cc5ab004724cf975914538ddd9aa65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of articles to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "articles not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/articles/{id}/view": {
            "get": {
                "tags": [
                    "articles"
                ],
                "summary": "Find articles by ID with their views",
                "description": "Returns a single articles with views",
                "operationId": "360c14692e503046fdb696d7f19e008d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of articles to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplier"
                    },
                    "404": {
                        "description": "articles not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    }
}